Add validateAggregatorSignature
