+++ beacon-chain/core/helpers/signing_root.go
@@ -1,9 +1,11 @@
 package helpers
 
 import (
+	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/go-ssz"
 	p2ppb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
 	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
+	"github.com/prysmaticlabs/prysm/shared/bls"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/params"
 )
@@ -13,6 +15,10 @@ const ForkVersionByteLength = 4
 
 // DomainByteLength length of domain byte array.
 const DomainByteLength = 4
+
+// ErrSigFailedToVerify returns when a signature of a block object(ie attestation, slashing, exit... etc)
+// failed to verify.
+var ErrSigFailedToVerify = errors.New("signature did not verify")
 
 // ComputeSigningRoot computes the root of the object by calculating the root of the object domain tree.
 //
@@ -36,6 +42,26 @@ func ComputeSigningRoot(object interface{}, domain []byte) ([32]byte, error) {
 		Domain:     domain,
 	}
 	return ssz.HashTreeRoot(container)
+}
+
+// VerifySigningRoot verifies the signing root of an object given it's public key, signature and domain.
+func VerifySigningRoot(obj interface{}, pub []byte, signature []byte, domain []byte) error {
+	publicKey, err := bls.PublicKeyFromBytes(pub)
+	if err != nil {
+		return errors.Wrap(err, "could not convert bytes to public key")
+	}
+	sig, err := bls.SignatureFromBytes(signature)
+	if err != nil {
+		return errors.Wrap(err, "could not convert bytes to signature")
+	}
+	root, err := ComputeSigningRoot(obj, domain)
+	if err != nil {
+		return errors.Wrap(err, "could not compute signing root")
+	}
+	if !sig.Verify(root[:], publicKey) {
+		return ErrSigFailedToVerify
+	}
+	return nil
 }
 
 // ComputeDomain returns the domain version for BLS private key to sign and verify with a zeroed 4-byte
